@@ -1,10 +1,10 @@
 package com.example.ben.smarttext;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.TypeConverters;
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+import androidx.room.TypeConverters;
 
-@Database(entities = {TextMessage.class}, version = 2)
+@Database(entities = {TextMessage.class}, version = 3)
 @TypeConverters({DateTypeConverer.class, UUIDTypeConverter.class})
 public abstract class AppDatabase extends RoomDatabase{
     public abstract TextMessageDAO getTextMessageDAO();
